chore: split out slash commands into multiple files Breakup the monolithic file don't register slashcommands on startup
