+++ dGame/dUtilities/CMakeLists.txt
@@ -8,9 +8,15 @@ set(DGAME_DUTILITIES_SOURCES "BrickDatabase.cpp"
 	"SlashCommandHandler.cpp"
 	"VanityUtilities.cpp")
 
+add_subdirectory(SlashCommands)
+
+foreach(file ${DGAME_DUTILITIES_SLASHCOMMANDS})
+	set(DGAME_DUTILITIES_SOURCES ${DGAME_DUTILITIES_SOURCES} "SlashCommands/${file}")
+endforeach()
+
 add_library(dUtilities OBJECT ${DGAME_DUTILITIES_SOURCES})
 target_precompile_headers(dUtilities REUSE_FROM dGameBase)
-target_include_directories(dUtilities PUBLIC "."
+target_include_directories(dUtilities PUBLIC "." "SlashCommands"
 	PRIVATE
 	"${PROJECT_SOURCE_DIR}/dGame/dComponents"
 	"${PROJECT_SOURCE_DIR}/dGame/dInventory" # transitive via PossessableComponent.h
