+++ dGame/dUtilities/SlashCommands/CMakeLists.txt
@@ -1,0 +1,6 @@
+set(DGAME_DUTILITIES_SLASHCOMMANDS
+	"DEVGMCommands.cpp"
+	"GMGreaterThanZeroCommands.cpp"
+	"GMZeroCommands.cpp"
+	PARENT_SCOPE
+)
