+++ dWorldServer/WorldServer.cpp
@@ -314,8 +314,8 @@ int main(int argc, char** argv) {
 	uint32_t sqlPingTime = 10 * 60 * currentFramerate; // 10 minutes in frames
 	uint32_t emptyShutdownTime = (cloneID == 0 ? 30 : 5) * 60 * currentFramerate; // 30 minutes for main worlds, 5 for all others.
 
-	// Register slash commands
-	SlashCommandHandler::Startup();
+	// Register slash commands is not in zone 0
+	if (zoneID != 0) SlashCommandHandler::Startup();
 
 	Game::logger->Flush(); // once immediately before the main loop
 	while (true) {
